0.7.10.10: 	Working INLINE inside MACROLET for user code. 		(more or less as per CSR sbcl-devel 2002-12-03) 	... change tricky cross-compiled inline functions to source 		transforms (including adding some functions to the 		function database) 	... use now-working FUNCTION-LAMBDA-EXPRESSION to get the 		expander function for local macros 	... build a complex LAMBDA-WITH-LEXENV for inlining user code 		(like CMUCL's INLINE-SYNTACTIC-CLOSURE-LAMBDA does) 	Some miscellaneous frobs 	... actually signal a style warning for array code (though it 		might actually be in currently-dead-but-should-be-alive 		code) 	... actually test the return value in the second half of 		filesys.test.sh 	... minor text adjustments (no more calling this period "early 		0.7.x")

+++ tests/filesys.test.sh
@@ -184,7 +184,12 @@ Lisp filename syntax idiosyncrasies)."
   (need-match "animal/vertebrate/mammal/../**/robot/*.*" nil)
   (need-match "animal/vertebrate/mammal/robot/../**/../**/*.*" nil))
 (need-matches)
+(sb-ext:quit :unix-status 52)
 EOF
+if [ $? != 52 ]; then
+    echo DIRECTORY/TRUENAME test part 1 failed, unexpected SBCL return code=$?
+    exit 1
+fi
 cd ..
 rm -r $testdir
